+++ node/src/types/kernel/api/block/insertBlock/payload.d.ts
@@ -15,12 +15,33 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+/**
+ * Insert blocks
+ */
+export interface IPayload {
+    /**
+     * block data content
+     */
+    data: string;
+    /**
+     * block data format
+     */
+    dataType: DataType;
+    /**
+     * block ID: insert before this block
+     */
+    nextID?: string;
+    /**
+     * block ID: insert into this block
+     */
+    parentID?: string;
+    /**
+     * block ID: insert after this block
+     */
+    previousID?: string;
+}
 
-/* directories */
-export * as getBlockKramdown from "./getBlockKramdown";
-export * as getChildBlocks from "./getChildBlocks";
-export * as insertBlock from "./insertBlock";
-
-/* files */
-export * as getBlockBreadcrumb from "./getBlockBreadcrumb";
-export * as getDocInfo from "./getDocInfo";
+/**
+ * block data format
+ */
+export type DataType = "markdown" | "dom";