+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 添加 API `/api/asset/upload` | Add API `/api/asset/upload`
 - 添加 API `/api/block/getChildBlocks` | Add API `/api/block/getChildBlocks`
 - 添加 API `/api/block/getBlockKramdown` | Add API `/api/block/getBlockKramdown`
+- 添加 API `/api/block/insertBlock` | Add API `/api/block/insertBlock`
 
 ## v0.0.5 / 2023-06-30